only allow import repo git url startswith http(s):// and git://

+++ repo/views.py
@@ -940,11 +940,11 @@ def fulfill_gitrepo(username, reponame, auth_type, remote_git_url, remote_userna
             shutil.copytree(GIT_BARE_REPO_PATH, repo_path)
     git_daemon_export_ok_file_path = '%s/%s' % (repo_path, 'git-daemon-export-ok')
     if auth_type == 0:
-        if not os.path.exists(git_daemon_export_ok_file_path):
+        if os.path.exists(repo_path) and not os.path.exists(git_daemon_export_ok_file_path):
             with open(git_daemon_export_ok_file_path, 'w') as _file:
                 _file.close()
     else:
-        if os.path.exists(git_daemon_export_ok_file_path):
+        if os.path.exists(repo_path) and os.path.exists(git_daemon_export_ok_file_path):
             os.remove(git_daemon_export_ok_file_path)
 
 def get_common_repo_dict(request, repo, user_name, repo_name, refs):
